implement sub command for gensig
